# openapi/aurix-openfinance.yaml
openapi: 3.0.3
info:
  title: Aurix Open Finance Brasil API
  description: |
    APIs do Open Finance Brasil (BACEN) seguindo o padrão FAPI-Brazil.
    **Fase 1**: Contas, transações, saldos, consentimento, dados pessoais.
    **Fase 2**: Cartões de crédito, faturas, transações de cartão.
    **Fase 3**: Empréstimos, seguros, PIX.
  version: 1.0.0
  contact:
    name: Aurix Platform
    url: https://github.com/aurix-core-banking
servers:
  - url: http://localhost:8096
    description: Desenvolvimento local
...
components:
  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.openfinancebrasil.org.br/oauth/authorize
          tokenUrl: https://auth.openfinancebrasil.org.br/oauth/token
          scopes:
            openid: OpenID Connect
            consents: Gerenciar consentimentos
            accounts: Acessar dados de contas
            credit-cards: Acessar dados de cartões
            loans: Acessar dados de empréstimos
            insurance: Acessar dados de seguros
            pix: Acessar dados de PIX
            customers: Acessar dados pessoais
  schemas:
...
    CartaoResponse:
      type: object
      properties:
        cartaoId:
          type: string
        bandeira:
          type: string
        finalNumero:
          type: string
        statusCartao:
          type: string
        limiteCredito:
          type: number
        limiteDisponivel:
          type: number
        valorFaturaAtual:
          type: number
        dataVencimentoFatura:
          type: string
          format: date
    FaturaResponse:
      type: object
      properties:
        faturaId:
          type: string
        valorTotal:
          type: number
        valorMinimo:
          type: number
        valorPago:
          type: number
        dataVencimento:
          type: string
          format: date
        dataPagamento:
          type: string
          format: date
    TransacaoCartaoResponse:
      type: object
      properties:
        transactionId:
          type: string
        valor:
          type: number
        moeda:
          type: string
        estabelecimento:
          type: string
        tipoTransacao:
          type: string
        dataTransacao:
          type: string
          format: date-time
    EmprestimoResponse:
      type: object
      properties:
        emprestimoId:
          type: string
        tipoEmprestimo:
          type: string
        valorContratado:
          type: number
        valorSaldoDevedor:
          type: number
        taxaJuros:
          type: number
        prazoMeses:
          type: integer
        parcelasPagas:
          type: integer
        parcelasRestantes:
          type: integer
        valorParcela:
          type: number
        dataContratacao:
          type: string
          format: date
        dataVencimentoPrimeiraParcela:
          type: string
          format: date
        statusEmprestimo:
          type: string
    SeguroResponse:
      type: object
      properties:
        apoliceId:
          type: string
        tipoSeguro:
          type: string
        nomeSeguradora:
          type: string
        premioMensal:
          type: number
        premioTotal:
          type: number
        valorSegurado:
          type: number
        dataInicio:
          type: string
          format: date
        dataFim:
          type: string
          format: date
        statusApolice:
          type: string
    PixResponse:
      type: object
      properties:
        pixId:
          type: string
        tipoPix:
          type: string
        chavedePix:
          type: string
        tipoChave:
          type: string
        valor:
          type: number
        moeda:
          type: string
        descricao:
          type: string
        statusPix:
          type: string
        dataPix:
          type: string
          format: date-time
...
security:
  - OAuth2: []
paths:
...
  /open-finance/v1/credit-cards:
    get:
      tags: [Cartões de Crédito]
      summary: Listar cartões de crédito
      operationId: listarCartoes
      parameters:
        - $ref: '#/components/parameters/FapiAuthDate'
        - $ref: '#/components/parameters/FapiIpAddress'
        - $ref: '#/components/parameters/FapiInteractionId'
      responses:
        '200':
          description: Lista de cartões
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CartaoResponse'
  /open-finance/v1/credit-cards/{creditCardId}:
    get:
      tags: [Cartões de Crédito]
      summary: Buscar cartão de crédito
      operationId: buscarCartao
      parameters:
        - name: creditCardId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/FapiAuthDate'
        - $ref: '#/components/parameters/FapiIpAddress'
        - $ref: '#/components/parameters/FapiInteractionId'
      responses:
        '200':
          description: Cartão encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartaoResponse'
        '404':
          description: Não encontrado
  /open-finance/v1/credit-cards/{creditCardId}/bills:
    get:
      tags: [Faturas]
      summary: Listar faturas do cartão
      operationId: listarFaturas
      parameters:
        - name: creditCardId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/FapiAuthDate'
        - $ref: '#/components/parameters/FapiIpAddress'
        - $ref: '#/components/parameters/FapiInteractionId'
      responses:
        '200':
          description: Lista de faturas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FaturaResponse'
  /open-finance/v1/credit-cards/{creditCardId}/bills/{billId}:
    get:
      tags: [Faturas]
      summary: Buscar fatura do cartão
      operationId: buscarFatura
      parameters:
        - name: creditCardId
          in: path
          required: true
          schema:
            type: string
        - name: billId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/FapiAuthDate'
        - $ref: '#/components/parameters/FapiIpAddress'
        - $ref: '#/components/parameters/FapiInteractionId'
      responses:
        '200':
          description: Fatura encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FaturaResponse'
        '404':
          description: Não encontrada
  /open-finance/v1/credit-cards/{creditCardId}/transactions:
    get:
      tags: [Transações de Cartão]
      summary: Listar transações do cartão
      operationId: listarTransacoesCartao
      parameters:
        - name: creditCardId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/FapiAuthDate'
        - $ref: '#/components/parameters/FapiIpAddress'
        - $ref: '#/components/parameters/FapiInteractionId'
      responses:
        '200':
          description: Lista de transações do cartão
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransacaoCartaoResponse'

  # ============================================================
  # Fase 3 — Empréstimos
  # ============================================================
  /open-finance/v1/loans:
    get:
      tags: [Empréstimos]
      summary: Listar empréstimos do cliente
      operationId: listarEmprestimos
      parameters:
        - $ref: '#/components/parameters/FapiAuthDate'
        - $ref: '#/components/parameters/FapiIpAddress'
        - $ref: '#/components/parameters/FapiInteractionId'
      responses:
        '200':
          description: Lista de empréstimos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmprestimoResponse'
  /open-finance/v1/loans/{loanId}:
    get:
      tags: [Empréstimos]
      summary: Buscar empréstimo
      operationId: buscarEmprestimo
      parameters:
        - name: loanId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/FapiAuthDate'
        - $ref: '#/components/parameters/FapiIpAddress'
        - $ref: '#/components/parameters/FapiInteractionId'
      responses:
        '200':
          description: Empréstimo encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmprestimoResponse'
        '404':
          description: Não encontrado

  # ============================================================
  # Fase 3 — Seguros
  # ============================================================
  /open-finance/v1/insurance:
    get:
      tags: [Seguros]
      summary: Listar seguros do cliente
      operationId: listarSeguros
      parameters:
        - $ref: '#/components/parameters/FapiAuthDate'
        - $ref: '#/components/parameters/FapiIpAddress'
        - $ref: '#/components/parameters/FapiInteractionId'
      responses:
        '200':
          description: Lista de seguros
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SeguroResponse'
  /open-finance/v1/insurance/{insuranceId}:
    get:
      tags: [Seguros]
      summary: Buscar seguro
      operationId: buscarSeguro
      parameters:
        - name: insuranceId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/FapiAuthDate'
        - $ref: '#/components/parameters/FapiIpAddress'
        - $ref: '#/components/parameters/FapiInteractionId'
      responses:
        '200':
          description: Seguro encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeguroResponse'
        '404':
          description: Não encontrado

  # ============================================================
  # Fase 3 — PIX
  # ============================================================
  /open-finance/v1/pix/credit-transfers:
    get:
      tags: [PIX]
      summary: Listar transferências PIX
      operationId: listarPix
      parameters:
        - $ref: '#/components/parameters/FapiAuthDate'
        - $ref: '#/components/parameters/FapiIpAddress'
        - $ref: '#/components/parameters/FapiInteractionId'
      responses:
        '200':
          description: Lista de PIX
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PixResponse'
  /open-finance/v1/pix/credit-transfers/{pixId}:
    get:
      tags: [PIX]
      summary: Buscar transferência PIX
      operationId: buscarPix
      parameters:
        - name: pixId
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/FapiAuthDate'
        - $ref: '#/components/parameters/FapiIpAddress'
        - $ref: '#/components/parameters/FapiInteractionId'
      responses:
        '200':
          description: PIX encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PixResponse'
        '404':
          description: Não encontrado
components:
  parameters:
    FapiAuthDate:
      name: x-fapi-auth-date
      in: header
      required: true
      schema:
        type: string
        format: date-time
    FapiIpAddress:
      name: x-fapi-auth-ip-address
      in: header
      required: true
      schema:
        type: string
    FapiInteractionId:
      name: x-fapi-interaction-id
      in: header
      required: true
      schema:
        type: string
        format: uuid